backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
django-cors-headers==3.14.0
Pillow==10.0.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.1